реализованы все методы нулевого этапа. добавлено возвращение StatusCode

+++ src/Application/AppServices/IRepository/IAnimalRepository.cs
@@ -12,7 +12,7 @@ namespace AppServices.IRepository
 {
     public interface IAnimalRepository
     {
-        public Task<Animal> GetAnimalById(int id);
+        public Task<Animal> GetAnimalById(long id);
 
         public Task<IReadOnlyCollection<Animal>> GetAllAnimal();
