+++ src/Application/AppServices/Services/Animal/IAnimalService.cs
@@ -9,7 +9,7 @@ namespace AppServices.Services.Animal
 {
     public interface IAnimalService
     {
-        public Task<InfoAnimalResponse> GetAnimalById(int id);
+        public Task<InfoAnimalResponse> GetAnimalById(long id);
 
         public Task<IReadOnlyCollection<InfoAnimalResponse>> GetAnimalByFillters(DateTime startDateTime, DateTime endDateTime, int chipperId, long chippingLocationId, string lifeStatus, string gender, int from, int size);
